src/test/java/cn/edu/zjnu/acm/judge/service/DeleteServiceTest.java
/*
 * Copyright 2018 ZJNU ACM.
 *
 * Licensed under the Apache License, Version 2.0 (the "License");
 * you may not use this file except in compliance with the License.
 * You may obtain a copy of the License at
 *
 *      http://www.apache.org/licenses/LICENSE-2.0
 *
 * Unless required by applicable law or agreed to in writing, software
 * distributed under the License is distributed on an "AS IS" BASIS,
 * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 * See the License for the specific language governing permissions and
 * limitations under the License.
 */
package cn.edu.zjnu.acm.judge.service;

import cn.edu.zjnu.acm.judge.Application;
import java.io.IOException;
import java.nio.file.Files;
import java.nio.file.Path;
import java.nio.file.Paths;
import java.util.concurrent.ScheduledFuture;
import java.util.concurrent.TimeUnit;
import lombok.extern.slf4j.Slf4j;
import org.junit.Test;
import org.junit.runner.RunWith;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.boot.test.context.SpringBootTest;
import org.springframework.test.context.junit4.SpringRunner;
import org.springframework.test.context.web.WebAppConfiguration;
import org.springframework.transaction.annotation.Transactional;

import static org.junit.Assert.assertTrue;

/**
 *
 * @author zhanhb
 */
@RunWith(SpringRunner.class)
@Slf4j
@SpringBootTest(classes = Application.class)
@Transactional
@WebAppConfiguration
public class DeleteServiceTest {

    @Autowired
    private DeleteService instance;

    /**
     * Test of delete method, of class DeleteService.
     */
    @Test
    public void testDelete() throws IOException, InterruptedException {
        log.info("delete");
        Path path = Files.createTempDirectory(Paths.get("target"), "tmp");
        Path txt = Files.createFile(path.resolve("a.txt"));
        ScheduledFuture<?> future = instance.delete(path);
        while (!future.isCancelled()) {
            TimeUnit.MILLISECONDS.sleep(200);
        }
        assertTrue(!Files.exists(txt));
    }

}
